+++ unionj-generator-backend/src/main/java/cloud/unionj/generator/backend/springboot/OutputType.java
@@ -1,0 +1,19 @@
+package cloud.unionj.generator.backend.springboot;
+
+/**
+ * @author created by tqccc
+ * @version v0.0.1
+ * description: cloud.unionj.generator.backend.springboot
+ * date:2021/1/17
+ */
+public enum OutputType {
+  /**
+   * check
+   */
+  CHECK,
+
+  /**
+   * overwrite
+   */
+  OVERWRITE
+}
